not drop schema if excludedDoctrineTables is not empty

+++ Services/DatabaseTools/ORMDatabaseTool.php
@@ -131,9 +131,15 @@ class ORMDatabaseTool extends AbstractDatabaseTool
 
         // TODO: handle case when using persistent connections. Fail loudly?
         $schemaTool = new SchemaTool($this->om);
-        $schemaTool->dropDatabase();
-        if (!empty($this->getMetadatas())) {
-            $schemaTool->createSchema($this->getMetadatas());
+        if (count($this->excludedDoctrineTables) > 0) {
+            if (!empty($this->getMetadatas())) {
+                $schemaTool->updateSchema($this->getMetadatas());
+            }
+        } else {
+            $schemaTool->dropDatabase();
+            if (!empty($this->getMetadatas())) {
+                $schemaTool->createSchema($this->getMetadatas());
+            }
         }
         $this->webTestCase->postFixtureSetup();
